perf: optimize core pipeline + add per-stage timing
Performance optimizations:
- normalize_image: fuse scale/mean/std into single multiply+add (减少 3 次数组操作)
- preprocess: remove redundant .astype(float32) after normalize_image
- perspective_transform: cache dst array to avoid repeated allocation
- order_points: avoid redundant .astype() when already float32
- text_detector: use math.hypot instead of np.linalg.norm for 2D distance
- postprocess: vectorized NMS (batch IoU instead of O(n²) Python loop)
- postprocess: pre-compute scale_arr outside contour loop
- postprocess: remove scaling=1.0 identity ops in _unclip
- postprocess: use uint8 mask directly as bool index (no .astype copy)
- recognition: pre-filter CTC non-blank/non-repeat positions with numpy
- text_recognizer: remove redundant _to_probs min/sum/allclose per call

Timing additions:
- Info mode: angle time, detection time, recognition time, total time
- Debug mode: per-box crop time, per-line recognition time

+++ ascend_ocr/preprocess.py
@@ -106,7 +106,7 @@ def preprocess_for_detection(
 
     scale, mean, std = _get_norm_params(cfg)
     tensor = normalize_image(resized, scale=scale, mean=mean, std=std)
-    tensor = np.expand_dims(tensor, axis=0).astype(np.float32)
+    tensor = np.expand_dims(tensor, axis=0)
     return tensor, (scale_h, scale_w), (resize_h, resize_w)
 
 
@@ -125,7 +125,7 @@ def preprocess_for_classification(img: np.ndarray, cfg: ClsConfig) -> np.ndarray
     resized = cv2.resize(img, (target_w, target_h), interpolation=cv2.INTER_LINEAR)
     scale, mean, std = _get_norm_params(cfg)
     tensor = normalize_image(resized, scale=scale, mean=mean, std=std)
-    return np.expand_dims(tensor, axis=0).astype(np.float32)
+    return np.expand_dims(tensor, axis=0)
 
 
 def preprocess_for_recognition(
@@ -188,7 +188,7 @@ def preprocess_for_recognition(
 
     scale, mean, std = _get_norm_params(cfg)
     tensor = normalize_image(resized, scale=scale, mean=mean, std=std)
-    tensor = np.expand_dims(tensor, axis=0).astype(np.float32)
+    tensor = np.expand_dims(tensor, axis=0)
     logger.debug(
         "Rec preprocess: crop %dx%d -> tensor %s (mode=%s)",
         w, h, tensor.shape, cfg.resize_mode,
